+++ src/main/groovy/fr/nereide/completion/provider/common/EntityFieldCompletionProvider.groovy
@@ -37,34 +37,19 @@ import fr.nereide.dom.element.entitymodel.ViewEntity
 import fr.nereide.project.OfbizProjectHelper
 import fr.nereide.project.PluginActivator
 import fr.nereide.project.pattern.OfbizPluginConstants
+import fr.nereide.project.utils.MiscUtils
 import fr.nereide.project.utils.PsiUtils
 import org.jetbrains.annotations.NotNull
 import org.jetbrains.plugins.groovy.lang.psi.api.statements.GrVariable
 import org.jetbrains.plugins.groovy.lang.psi.api.statements.expressions.GrReferenceExpression
 import org.jetbrains.plugins.groovy.lang.psi.dataFlow.types.TypeInferenceHelper
 
-import java.util.regex.Matcher
-import java.util.regex.Pattern
-
 /**
  * Part of the OFBiz plugin completion system
  */
 abstract class EntityFieldCompletionProvider extends CompletionProvider<CompletionParameters> {
 
-    private static final Pattern ENTITY_NAME_PATTERN = Pattern.compile("(['\"](.*?)['\"])")
     public static final String DBLE_QUOTE = '\"'
-
-    /**
-     * Extracts entity name from declarations like
-     * <code> EntityQuery.use(delegator).from() </code>
-     * @param declaration the declaration string
-     * @return the entity name
-     */
-    static String getEntityNameFromDeclarationString(String declaration) {
-        Matcher matcher = ENTITY_NAME_PATTERN.matcher(declaration)
-        String entityName = matcher.find() ? matcher.group(0) : null
-        return entityName ? entityName.substring(1, entityName.length() - 1) : null
-    }
 
     /**
      * Add lookup element to display, with the fields and the entity name at the end of the line
@@ -86,7 +71,7 @@ abstract class EntityFieldCompletionProvider extends CompletionProvider<Completi
     static String getEntityNameFromQuery(PsiElement element, Class methodTypeClass) {
         PsiElement originMethod = getParentOfType(element, methodTypeClass)
         if (originMethod && originMethod.text.contains(OfbizPluginConstants.QUERY_FROM_STATEMENT)) {
-            return getEntityNameFromDeclarationString(originMethod.text)
+            return MiscUtils.getEntityNameFromDeclarationString(originMethod.text)
         }
         return null
     }
@@ -117,7 +102,7 @@ abstract class EntityFieldCompletionProvider extends CompletionProvider<Completi
         }.toList()?.last()
         if (!lastQuery) return ''
         PsiElement lastAssignExpr = getParentOfType(lastQuery.element, assigmentClass)
-        return getEntityNameFromDeclarationString(getAssigmentString(lastAssignExpr))
+        return MiscUtils.getEntityNameFromDeclarationString(getAssigmentString(lastAssignExpr))
     }
 
     /**