+++ src/main/groovy/fr/nereide/completion/provider/groovy/GroovyEntityFieldCompletionProvider.groovy
@@ -26,6 +26,7 @@ import com.intellij.psi.PsiMethod
 import com.intellij.psi.PsiVariable
 import fr.nereide.completion.provider.common.EntityFieldCompletionProvider
 import fr.nereide.project.pattern.OfbizGroovyPatterns
+import fr.nereide.project.utils.MiscUtils
 import org.jetbrains.plugins.groovy.lang.psi.api.statements.GrForStatement
 import org.jetbrains.plugins.groovy.lang.psi.api.statements.GrLoopStatement
 import org.jetbrains.plugins.groovy.lang.psi.api.statements.GrVariable
@@ -79,7 +80,7 @@ abstract class GroovyEntityFieldCompletionProvider extends EntityFieldCompletion
         PsiExpression init = initialElement.initializer
         String declarationString = init ? init.text : initialElement.initializerGroovy?.text
         if (declarationString && !(declarationString == 'null')) {
-            return getEntityNameFromDeclarationString(declarationString)
+            return MiscUtils.getEntityNameFromDeclarationString(declarationString)
         }
         GrLoopStatement oldFashionedLoop = getParentOfType(initialElement, GrLoopStatement)
         if (oldFashionedLoop) {